REFACTOR: Remove excess complexity with AVM indexer.
